refactor: move BuilderEventView to it's own file
